Add in inversion filter to svg lib so we can mimic how postscript
The postscript library offers an inversion function that places a
solid black rectangle behind the image and then inverts the
colors (black and white) you are using and draws the current image
using this inversion. The svg lib now creates a top-level filter with
every image so that it can use that filter to create a similar
inversion effect.