Redo ch 10 interpreter from Little Schemer
To make sure I understand the interpreter from chapter 10 before
I embark on chapter 20 of the seasoned schemer, I redid it and
tried to fully understand the mechanics of the interpreter. I
played around with extending it a bit to allow for recursive
functions, but I haven't thought the scope implications of my
change so it is commented out for now.